make it possible to disable legacy comms
not enabled anywhere yet, but can be tested by adding -DDISABLE_LEGACY_COMMS to compiler flags
